// CloudIMS/Cloud-IMS-App/ClientApp/src/app/classes/inventory-management/InventoryIn/IInventoryInTrxHeader.interface.ts
export interface IInventoryInTrxHeader {
  transactionNo: string;
  transactionDate: Date;
  receivedDate: Date;
  receivedBy: string;
  poNumber: string;
  invoiceNo: string;
  referenceNo: string;
  documnetNo: string;
  supplier: string;
  remarks: string;
}

export interface IListOfUserAccount {
  userID: string;
  userName: string;
}